monix-reactive partial fix

+++ monix-reactive/shared/src/test/scala/monix/reactive/consumers/TransformInputConsumerSuite.scala
@@ -35,7 +35,7 @@ object TransformInputConsumerSuite extends BaseTestSuite {
     check1 { (random: Observable[Int]) =>
       val source = random.map(Math.floorMod(_, 10))
       val consumer = Consumer.foldLeft[Long, Long](0L)(_ + _)
-      val transformed = consumer.transformInput[Int](_.map(_ + 100))
+      val transformed = consumer.transformInput[Int](_.map(_.toLong + 100))
       source.consumeWith(transformed) <-> source.foldLeftL(0L)(_ + _ + 100)
     }
   }
