+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/CacheSuite.scala
@@ -25,7 +25,7 @@ import scala.util.Success
 object CacheSuite extends BaseOperatorSuite {
   def createObservable(c: Int) = Some {
     val o = Observable.range(0, c).cache
-    Sample(o, c, c * (c - 1) / 2, Zero, Zero)
+    Sample(o, c, (c * (c - 1) / 2).toLong, Zero, Zero)
   }
 
   def observableInError(sourceCount: Int, ex: Throwable) = None
