+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/CombineLatest4Suite.scala
@@ -30,7 +30,7 @@ object CombineLatest4Suite extends BaseOperatorSuite {
     val o1 = Observable.now(1)
     val o2 = Observable.now(2)
     val o3 = Observable.now(3)
-    val o4 = Observable.range(0, sourceCount)
+    val o4 = Observable.range(0L, sourceCount.toLong)
     val o = Observable.combineLatestMap4(o1, o2, o3, o4)(_ + _ + _ + _)
 
     Sample(o, count(sourceCount), sum(sourceCount), waitFirst, waitNext)
@@ -45,7 +45,7 @@ object CombineLatest4Suite extends BaseOperatorSuite {
     val o1 = Observable.now(1)
     val o2 = Observable.now(2)
     val o3 = Observable.now(3)
-    val flawed = createObservableEndingInError(Observable.range(0, sourceCount), ex)
+    val flawed = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex)
     val o = Observable.combineLatestMap4(o1, o2, o3, flawed)(_ + _ + _ + _)
 
     Sample(o, count(sourceCount - 1), sum(sourceCount - 1), waitFirst, waitNext)
@@ -56,7 +56,7 @@ object CombineLatest4Suite extends BaseOperatorSuite {
     val o1 = Observable.now(1)
     val o2 = Observable.now(2)
     val o3 = Observable.now(3)
-    val o4 = Observable.range(0, sourceCount)
+    val o4 = Observable.range(0L, sourceCount.toLong)
     val o = Observable.combineLatestMap4(o1, o2, o3, o4) { (a1, a2, a3, a4) =>
       if (a4 == sourceCount - 1) throw dummy else a1 + a2 + a3 + a4
     }
