+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DebounceRepeatedSuite.scala
@@ -26,7 +26,7 @@ object DebounceRepeatedSuite extends BaseOperatorSuite {
       .now(1L)
       .delayOnComplete(1.day)
       .debounceRepeated(1.second)
-      .take(sourceCount)
+      .take(sourceCount.toLong)
 
     val count = sourceCount
     val sum = sourceCount
