+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DropLastSuite.scala
@@ -25,7 +25,7 @@ object DropLastSuite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = {
     require(sourceCount > 0, "sourceCount should be strictly positive")
     Some {
-      val o = Observable.range(0, sourceCount).dropLast(10)
+      val o = Observable.range(0L, sourceCount.toLong).dropLast(10)
       Sample(o, count(sourceCount - 10), sum(sourceCount - 10), Zero, Zero)
     }
   }
@@ -40,7 +40,7 @@ object DropLastSuite extends BaseOperatorSuite {
   def observableInError(sourceCount: Int, ex: Throwable) = {
     require(sourceCount > 0, "sourceCount should be strictly positive")
     Some {
-      val o = Observable.range(0, sourceCount).endWithError(ex).dropLast(1)
+      val o = Observable.range(0L, sourceCount.toLong).endWithError(ex).dropLast(1)
       Sample(o, count(sourceCount - 1), sum(sourceCount - 1), Zero, Zero)
     }
   }
