+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/FlatScanDelayErrorSuite.scala
@@ -31,8 +31,8 @@ object FlatScanDelayErrorSuite extends BaseOperatorSuite {
 
   def create(sourceCount: Int, ex: Throwable = null) = Some {
     val source =
-      if (ex == null) Observable.range(0, sourceCount)
-      else Observable.range(0, sourceCount).endWithError(ex)
+      if (ex == null) Observable.range(0L, sourceCount.toLong)
+      else Observable.range(0L, sourceCount.toLong).endWithError(ex)
 
     val o = source.flatScanDelayErrors(1L)((acc, elem) =>
       Observable
