+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/IntersperseSuite.scala
@@ -25,7 +25,7 @@ import scala.concurrent.duration._
 
 object IntersperseSuite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Option {
-    val sample = Observable.range(0, sourceCount).intersperse(1L)
+    val sample = Observable.range(0L, sourceCount.toLong).intersperse(1L)
     Sample(sample, count(sourceCount), sum(sourceCount), 0.seconds, 0.seconds)
   }
 
@@ -33,7 +33,7 @@ object IntersperseSuite extends BaseOperatorSuite {
   def sum(sourceCount: Int) = (sourceCount * (sourceCount - 1)) / 2 + sourceCount - 1
 
   def observableInError(sourceCount: Int, ex: Throwable) = Some {
-    val sample = createObservableEndingInError(Observable.range(0, sourceCount), ex).intersperse(0L)
+    val sample = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex).intersperse(0L)
     Sample(sample, count(sourceCount), sum(sourceCount), 0.seconds, 0.seconds)
   }
