+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/MapAccumulateSuite.scala
@@ -37,7 +37,7 @@ object MapAccumulateSuite extends BaseOperatorSuite {
     sourceCount
 
   def observableInError(sourceCount: Int, ex: Throwable) = Some {
-    val o = createObservableEndingInError(Observable.range(0, sourceCount), ex)
+    val o = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex)
       .mapAccumulate(0L)((acc, elem) => (acc + elem, acc * elem))
 
     Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
@@ -50,7 +50,7 @@ object MapAccumulateSuite extends BaseOperatorSuite {
   }
 
   def brokenUserCodeObservable(sourceCount: Int, ex: Throwable) = Some {
-    val o = Observable.range(0, sourceCount).mapAccumulate(0L) { (acc, elem) =>
+    val o = Observable.range(0L, sourceCount.toLong).mapAccumulate(0L) { (acc, elem) =>
       if (elem == sourceCount - 1)
         throw ex
       else
