+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/MapParallelOrderedSuite.scala
@@ -34,7 +34,7 @@ object MapParallelOrderedSuite extends BaseOperatorSuite {
 
   def createObservable(sourceCount: Int): Some[MapParallelOrderedSuite.Sample] = {
     Some {
-      val o = Observable.range(0, sourceCount).mapParallelOrdered(parallelism = 4)(x => Task(x))
+      val o = Observable.range(0L, sourceCount.toLong).mapParallelOrdered(parallelism = 4)(x => Task(x))
       Sample(o, count(sourceCount), sum(sourceCount), waitFirst, waitNext)
     }
   }
@@ -112,7 +112,7 @@ object MapParallelOrderedSuite extends BaseOperatorSuite {
       var received = 0
       var total = 0L
 
-      val obs = Observable.range(0, sourceCount).mapParallelOrdered(parallelism = 4)(x => Task.now(x))
+      val obs = Observable.range(0L, sourceCount.toLong).mapParallelOrdered(parallelism = 4)(x => Task.now(x))
       obs.unsafeSubscribeFn(new Observer[Long] {
         private[this] var sum = 0L
 
@@ -142,7 +142,7 @@ object MapParallelOrderedSuite extends BaseOperatorSuite {
       var received = 0
       var total = 0L
 
-      val obs = Observable.range(0, sourceCount).mapParallelOrdered(parallelism = 4)(x => Task(x))
+      val obs = Observable.range(0L, sourceCount.toLong).mapParallelOrdered(parallelism = 4)(x => Task(x))
       obs.unsafeSubscribeFn(new Observer[Long] {
         private[this] var sum = 0L
